requirements.txt
appdirs==1.4.4
blinker==1.4
boto==2.49.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.0
dnspython==2.0.0
email-validator==1.1.2
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-WTF==0.14.3
forex-python==1.5
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
simplejson==3.17.2
six==1.15.0
spotipy==2.18.0
source==1.2.0
urllib3==1.26.2
virtualenv==20.0.21
Werkzeug==1.0.1
WTForms==2.3.3
